// src/utils/enums/Sector.ts
export enum Sector {
  ADMINISTRACION = "Administracion",
  COORDINACION = "Coordinacion",
  PROFESIONALES = "Profesionales",
  CONTADURIA = "Contaduria",
  COMPRAS = "Compras",
  LIQUIDACIONES = "Liquidaciones",
  FACTURACION = "Facturacion",
  RRHH = "RRHH",
  RECLAMOS = "Reclamos",
  RECEPCION = "Recepcion",
  VARIABLE = "Variable",
}