// src/utils/enums/UserRole.ts
export enum UserRole {
  ADMINISTRADOR = "Administrador",
  COORDINACION = "Coordinacion",
  PROFESIONALES = "Profesionales",
  CONTADURIA = "Contaduria",
  COMPRAS = "Compras",
  LIQUIDACIONES = "Liquidaciones",
  COORDINADOR_SECTOR = "Coordinador de Sector",
  FACTURACION = "Facturacion",
  RECURSOS_HUMANOS = "Recursos Humanos",
  RECLAMOS = "Reclamos",
  RECEPCION = "Recepcion",
}